refactor(xpu): move INT8 MM weight to platform ops
