+++ lightx2v/common/ops/mm/mm_weight.py
@@ -2695,63 +2695,3 @@ class MMWeightFp8IntelXpu(MMWeightQuantTemplate):
         if squeeze_output:
             output = output.unsqueeze(0)
         return output
-
-
-@MM_WEIGHT_REGISTER("int8-intel-xpu")
-class MMWeightInt8IntelXpu(MMWeightQuantTemplate):
-    """Dynamic per-token W8A8 linear backed by the Intel XPU extension.
-
-    Checkpoint weights are signed INT8 in PyTorch Linear layout ``[N, K]``
-    with one FP32 dequantization scale per output channel. Activations remain
-    BF16/FP16 at the Python boundary and are quantized rowwise inside the SYCL
-    extension before the native oneDNN S8 x S8 matmul.
-    """
-
-    def __init__(
-        self,
-        weight_name,
-        bias_name,
-        create_cuda_buffer=False,
-        create_cpu_buffer=False,
-        lazy_load=False,
-        lazy_load_file=None,
-        is_post_adapter=False,
-        lora_prefix="diffusion_model.blocks",
-        lora_path="",
-    ):
-        super().__init__(
-            weight_name,
-            bias_name,
-            create_cuda_buffer,
-            create_cpu_buffer,
-            lazy_load,
-            lazy_load_file,
-            is_post_adapter,
-            lora_prefix,
-            lora_path,
-        )
-        self.load_func = self.load_int8_perchannel_sym
-        # The extension consumes the checkpoint's native Linear [N, K]
-        # layout. Other quant backends commonly transpose it to [K, N].
-        self.weight_need_transpose = False
-        self.scale_force_fp32 = True
-        self.bias_force_fp32 = True
-
-    def apply(self, input_tensor):
-        if sycl_kernels is None or not hasattr(sycl_kernels, "onednn_w8a8_int8"):
-            raise RuntimeError("int8-intel-xpu requires a lightx2v_kernel_xpu build that exports sycl_kernels.onednn_w8a8_int8")
-        if input_tensor.dtype not in (torch.float16, torch.bfloat16):
-            raise RuntimeError(f"int8-intel-xpu requires FP16/BF16 activations, got {input_tensor.dtype} for {self.weight_name}")
-
-        original_shape = input_tensor.shape[:-1]
-        input_2d = input_tensor.reshape(-1, input_tensor.shape[-1]).contiguous()
-        output = sycl_kernels.onednn_w8a8_int8(
-            input_2d,
-            self.weight.contiguous(),
-            self.weight_scale.reshape(-1).float().contiguous(),
-            self._get_actual_bias(),
-        )
-        output = output.reshape(*original_shape, self.weight.shape[0])
-        if self.has_lora_branch:
-            output = output + self.apply_lora(input_2d).reshape_as(output)
-        return output
